remove lesson9 from 351005 to 351001
